refactor axis rendering into separate component
